Restore setup.py only for purposes of dynamic versioning
